requirements.txt
Django==6.0.4
djangorestframework==3.14.0